requirements.txt
Click
GitPython
PyGithub
